Regroup tests to there own entity package
