+++ auth-service/src/test/java/com/petclinic/auth/User/AuthServiceUserServiceTests.java
@@ -1,4 +1,4 @@
-package com.petclinic.auth;
+package com.petclinic.auth.User;
 
 import com.petclinic.auth.Role.Role;
 import com.petclinic.auth.Role.RoleIDLessDTO;
@@ -49,10 +49,6 @@ public class AuthServiceUserServiceTests {
     @Autowired
     private UserService userService;
 
-
-    private UserController userController;
-    private final UserIDLessDTO ID_LESS_USER = new UserIDLessDTO();
-
     @BeforeEach
     void setup() {
         userRepo.deleteAllInBatch();
@@ -62,7 +58,9 @@ public class AuthServiceUserServiceTests {
     @DisplayName("Create new user")
     void create_new_user() {
         UserIDLessDTO userIDLessDTO = new UserIDLessDTO(USER, PASS, EMAIL);
+
         final User createdUser = userService.createUser(userIDLessDTO);
+
         assertEquals(createdUser.getUsername(), userIDLessDTO.getUsername());
         assertEquals(createdUser.getPassword(), userIDLessDTO.getPassword());
         assertEquals(createdUser.getEmail(), userIDLessDTO.getEmail());
@@ -74,8 +72,8 @@ public class AuthServiceUserServiceTests {
 
         final String CHANGE = "change";
         final User u = new User(USER, PASS, EMAIL);
-        userRepo.save(u);
 
+        userRepo.save(u);
         User user = userService.passwordReset(u.getId(), CHANGE);
 
         Optional<User> find = userRepo.findById(user.getId());
@@ -87,8 +85,8 @@ public class AuthServiceUserServiceTests {
     @DisplayName("Reset password, passed wrong ID")
     void test_user_password_reset_ID() {
 
-        final String CHANGE = "change";
         final User u = new User(USER, PASS, EMAIL);
+
         userRepo.save(u);
 
         assertThrows(NotFoundException.class, () -> userService.passwordReset(10000, ""));
@@ -99,7 +97,7 @@ public class AuthServiceUserServiceTests {
     void create_new_user_with_same_email() {
         UserIDLessDTO userIDLessDTO = new UserIDLessDTO(USER, PASS, EMAIL);
         User userMap = userMapper.idLessDTOToModel(userIDLessDTO);
-        User saved = userRepo.save(userMap);
+        userRepo.save(userMap);
         assertThrows(DuplicateKeyException.class, () -> userService.createUser(userIDLessDTO));
     }
 }
