--- auth-service/src/test/java/com/petclinic/auth/User/UserControllerTest.java
@@ -1,16 +1,0 @@
-package com.petclinic.auth.User;
-
-import org.junit.jupiter.api.Test;
-
-import static org.junit.jupiter.api.Assertions.*;
-
-class UserControllerTest {
-
-    @Test
-    void createUser() {
-    }
-
-    @Test
-    void passwordReset() {
-    }
-}
